taint/test/mochitest/test_location_sinks.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="utf-8">
    <title>Test Location Taint Sinks</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
    <script>

      let string_content = "hello";
      let sink_names = [
	  "location.hash",
	  "location.host",
	  "location.pathname",
	  "location.port",
	  "location.protocol",
	  "location.search",
	  "location.assign",
	  "location.href",
	  "location.replace"
      ];

      let i = 0;
      
      SimpleTest.waitForExplicitFinish();
      addEventListener("__taintreport", (report) => {
          let flow = report.detail.str.taint[0].flow;
          is(flow[1].operation, sink_names[i], "Check sink name");

	  i += 1;
	  if (i >= sink_names.length) {
              SimpleTest.finish();
	  }
      }, false);

      let taint_string = String.tainted(string_content);

      /// Location Sinks
      let url = location.href;
      url.untaint();
      console.error(url);

      location.hash = taint_string;
      location.host = String.tainted(location.host);
      location.pathname = String.tainted(location.pathname);
      try {
	  // Need to catch as will throw file not found error
	  location.port = String.tainted("80");
      } catch (error) {
	  console.log(error.msg);
      }
      location.protocol = String.tainted(location.protocol.untaint());
      location.search = taint_string;

      location.assign(String.tainted(url));
      location.href = String.tainted(url);
      location.replace(String.tainted(url));

    </script>
  </head>
  <body>
    <p id="display"></p>
    <div id="content" style="display: none"></div>
    <p id="test"></p>
  </body>
</html>
